# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /comercial/kpis:
    get:
      operationId: getComercialKpis
      tags: [comercial]
      summary: Commercial dashboard KPIs
      parameters:
        - { name: date_range, in: query, schema: { type: string, enum: [today, week, month, custom] } }
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
        - { name: vendedor, in: query, schema: { type: string } }
        - { name: estado, in: query, schema: { type: string } }
        - { name: cliente, in: query, schema: { type: string } }
      responses:
        "200":
          description: KPIs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ComercialKpis" }
  /comercial/reports:
    get:
      operationId: getComercialReports
      tags: [comercial]
      summary: Quotation opportunities table
      parameters:
        - { name: date_range, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
        - { name: vendedor, in: query, schema: { type: string } }
        - { name: estado, in: query, schema: { type: string } }
        - { name: cliente, in: query, schema: { type: string } }
        - { name: tecnico, in: query, schema: { type: string } }
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ComercialReportsResponse" }
  /comercial/report/{id}:
    patch:
      operationId: patchComercialReport
      tags: [comercial]
      summary: Update quotation status
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ComercialUpdateBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                required: [ok]
  /comercial/charts:
    get:
      operationId: getComercialCharts
      tags: [comercial]
      summary: All commercial chart datasets
      parameters:
        - { name: date_range, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
        - { name: vendedor, in: query, schema: { type: string } }
        - { name: estado, in: query, schema: { type: string } }
        - { name: cliente, in: query, schema: { type: string } }
      responses:
        "200":
          description: Chart data
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ComercialChartsData" }
  /comercial/vendedores:
    get:
      operationId: getComercialVendedores
      tags: [comercial]
      summary: Salesperson KPI table
      parameters:
        - { name: date_range, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
        - { name: estado, in: query, schema: { type: string } }
        - { name: cliente, in: query, schema: { type: string } }
      responses:
        "200":
          description: Vendedores list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ComercialVendedoresResponse" }
  /comercial/chat:
    post:
      operationId: postComercialChat
      tags: [comercial]
      summary: Commercial AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DashboardChatRequest" }
      responses:
        "200":
          description: AI answer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DashboardChatResponse" }
components:
  schemas:
...
    DashboardChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required:
        - role
        - content
    DashboardChatRequest:
      type: object
      properties:
        message:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/DashboardChatMessage"
      required:
        - message
    DashboardChatResponse:
      type: object
      properties:
        answer:
          type: string
      required:
        - answer
...
    ComercialKpis:
      type: object
      properties:
        total: { type: integer }
        enviadas: { type: integer }
        no_enviadas: { type: integer }
        pendientes: { type: integer }
        conversion_rate: { type: number }
        avg_response_times:
          type: array
          items:
            type: object
            properties:
              vendedor: { type: string }
              avg_hours: { type: number }
            required: [vendedor, avg_hours]
      required: [total, enviadas, no_enviadas, pendientes, conversion_rate, avg_response_times]
    ComercialReport:
      type: object
      properties:
        id: {}
        numero_reporte: {}
        created_at: { type: string, nullable: true }
        tecnico_nombre: { type: string, nullable: true }
        cliente_nombre: { type: string }
        trabajo_realizado: { type: string, nullable: true }
        descripcion_cotizacion: { type: string, nullable: true }
        cotizacion_estado: { type: string }
        cotizacion_enviada_at: { type: string, nullable: true }
        cotizacion_vendedor: { type: string, nullable: true }
        cotizacion_motivo_no_envio: { type: string, nullable: true }
        correo_comercial_enviado_at: { type: string, nullable: true }
        elapsed_hours: { type: number, nullable: true }
        response_hours: { type: number, nullable: true }
      required: [id, numero_reporte, cliente_nombre, cotizacion_estado]
    ComercialReportsResponse:
      type: object
      properties:
        reports:
          type: array
          items: { $ref: "#/components/schemas/ComercialReport" }
        total: { type: integer }
      required: [reports, total]
    ComercialUpdateBody:
      type: object
      properties:
        cotizacion_estado:
          type: string
          enum: [enviada, no_enviada]
        cotizacion_vendedor: { type: string }
        cotizacion_motivo_no_envio: { type: string }
      required: [cotizacion_estado]
    ComercialChartsData:
      type: object
      properties:
        porEstado:
          type: array
          items:
            type: object
            properties:
              estado: { type: string }
              value: { type: integer }
              color: { type: string }
            required: [estado, value, color]
        porVendedor:
          type: array
          items:
            type: object
            properties:
              nombre: { type: string }
              enviadas: { type: integer }
              no_enviadas: { type: integer }
              pendientes: { type: integer }
            required: [nombre, enviadas, no_enviadas, pendientes]
        tiempoVendedor:
          type: array
          items:
            type: object
            properties:
              nombre: { type: string }
              avg_horas: { type: number }
            required: [nombre, avg_horas]
        topClientes:
          type: array
          items:
            type: object
            properties:
              cliente: { type: string }
              oportunidades: { type: integer }
            required: [cliente, oportunidades]
      required: [porEstado, porVendedor, tiempoVendedor, topClientes]
    ComercialVendedor:
      type: object
      properties:
        nombre: { type: string }
        total: { type: integer }
        enviadas: { type: integer }
        no_enviadas: { type: integer }
        pendientes: { type: integer }
        tasa: { type: number }
        avg_horas: { type: number, nullable: true }
      required: [nombre, total, enviadas, no_enviadas, pendientes, tasa]
    ComercialVendedoresResponse:
      type: object
      properties:
        vendedores:
          type: array
          items: { $ref: "#/components/schemas/ComercialVendedor" }
      required: [vendedores]